requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# photon_bot/plugins/commands/ranking.py: 8
Pillow == 8.4.0

# photon_bot/plugins/commands/fun_games.py: 6
akinator.py == 5.0.0

# photon_bot/plugins/commands/web.py: 3
asyncpraw == 7.5.0

# photon_bot/utils.py: 4
mysql == 0.0.3

# photon_bot/plugins/commands/fun_basic.py: 6
nekos.py == 1.0.3

# photon_bot/plugins/commands/discord_info.py: 5
pygount == 1.3.0

# photon_bot/plugins/commands/utilities.py: 6
qrcode == 7.3.1

# photon_bot/plugins/commands/fun_basic.py: 7
# photon_bot/plugins/commands/ranking.py: 5
# photon_bot/plugins/commands/web.py: 6
requests == 2.25.1

# photon_bot/plugins/commands/web.py: 7
srcomapi == 0.3.3

# photon_bot/plugins/commands/web.py: 8
wikipedia == 1.4.0

# photon_bot/plugins/commands/web.py: 9
wolframalpha == 5.0.0
